refactor: Index 파일 활용
