+++ src/components/common/index.ts
@@ -1,0 +1,2 @@
+export { default as InputForm } from "./InputForm";
+export { default as Button } from "./Button";
